requirements.txt
# Core geospatial and Earth Engine
earthengine-api>=0.1.384
geemap>=0.30.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
rasterio>=1.3.0
pystac-client>=0.7.0
xarray>=2023.1.0

# Machine Learning - CPU/GPU compatible
scikit-learn>=1.3.0
torch>=2.0.0
tensorflow>=2.13.0  # For GPU support
keras>=2.13.0

# Deep Learning utilities
pytorch-lightning>=2.0.0
tensorboard>=2.13.0

# Spatial analysis
geopandas>=0.14.0
shapely>=2.0.0
pysal>=23.1
esda>=2.5.0  # For Getis-Ord Gi*
libpysal>=4.9.0

# Clustering
scikit-learn-extra>=0.3.0  # For additional clustering algorithms
hdbscan>=0.8.33

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
folium>=0.14.0
ipyleaflet>=0.17.0

# Workflow orchestration
prefect>=2.14.0
dask[complete]>=2023.1.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
joblib>=1.3.0

# Optional: For enhanced performance
numba>=0.57.0
bottleneck>=1.3.7
